+++ src/scripts/autoload/cf7-enable-shortcodes.php
@@ -1,0 +1,4 @@
+<?php
+
+// Activate Shortcode Execution for Contact Form 7
+add_filter( 'wpcf7_form_elements', 'do_shortcode' );