+++ src/scripts/autoload/plugin-dequeue-fontawesome.php
@@ -1,0 +1,9 @@
+<?php
+
+/**
+ * Disable load FontAwesome from CDN
+ */
+add_action( 'wp_enqueue_scripts', static function () {
+   wp_dequeue_style('font-awesome-official');
+   wp_deregister_style('font-awesome-official');
+}, 20);