+++ src/scripts/autoload/wp-disable-char-replacement.php
@@ -1,0 +1,8 @@
+<?php
+
+// Remove WP character replacement
+function astc_disable_wptexturize(): bool {
+   return false;
+}
+
+add_filter('run_wptexturize', 'astc_disable_wptexturize');