+++ src/scripts/autoload/wp-load-admin-styles.php
@@ -1,0 +1,9 @@
+<?php
+
+/**
+ * Enqueue styles in admin area
+ */
+function custom_admin_styles() {
+   wp_enqueue_style('custom-admin-styles', get_stylesheet_directory_uri() . '/admin/wp-admin.css');
+}
+add_action('admin_enqueue_scripts', 'custom_admin_styles');