+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Changelog
+
+# 1.4.0
+- Autoload PHP files
+- Performance improvements
+- Add auto theme switching without page reload
+
 # 1.3.8
 - Remove auto convert theme class